Restructure to new src & file naming conventions.
