+++ components/js-node.js
@@ -24,8 +24,8 @@ var _ = require('underscore');
 var fs = require('fs');
 var noflo = require('noflo');
 
-var basenode = require('./base-node');
-var basefnode = require('./base-file-node');
+var basenode = require('../src/base-node');
+var basefnode = require('../src/base-file-node');
 
 exports.getComponent = function() {
   return _.extend(
