+++ src/base-node.js
