+++ src/event-component-factory.js
