+++ src/promise-component.js
